chore: renamed inter-tx to icaauth
