+++ proto/akash/icaauth/tx.proto
@@ -1,13 +1,13 @@
 syntax = "proto3";
 
-package akash.intertx;
+package akash.icaauth;
 
-option go_package = "github.com/ovrclk/akash/x/inter-tx/types";
+option go_package = "github.com/ovrclk/akash/x/icaauth/types";
 
 import "gogoproto/gogo.proto";
 import "google/protobuf/any.proto";
 
-// Msg defines the intertx Msg service.
+// Msg defines the icaauth Msg service.
 service Msg {
   // Register defines a rpc handler for MsgRegisterAccount
   rpc RegisterAccount(MsgRegisterAccount) returns (MsgRegisterAccountResponse);
